requirements.txt
PyTorch
Plotly
scikit-learn
NumPy
pandas